requirements.txt
farm-haystack==0.8.0
tornado==6.1
ray==1.9.2
memory-profiler==0.60.0